feat: simplify everything

+++ ScannerCore/Humanize.cs
@@ -10,7 +10,7 @@ namespace ScannerCore
     {
         public static string Size(long size)
         {
-            Single output = size;
+            double output = size;
             int sufIdx = 0;
             while (Math.Abs(output) > 1024 && sufIdx < Suffixes.Length - 1)
             {
